Issue #16: Set up Elastic Beanstalk integration
## Problem
Elastic Beanstalk will be used to deploy the application. The service provides the ability to run Docker images and
connect the Docker container via the Nginx reverse proxy. This provides an isolated container that is fully in control
while taking minimal dependencies on the Elastic Beanstalk environment except to run Docker.

## Solution
* Add Elastic Beanstalk extensions that configure the environment including ngnix
* Fix bug in Auth0 setting code
* Update a few issues in the README.md file

## Test
* Successfully deployed to my AWS account after creating an RDS instance and DB to use.
* Add unit tests for the Auth0 setting code to ensure no future regression